% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_riskprogram_table.R
\name{create_risk_prog}
\alias{create_risk_prog}
\title{Combine Organization and Division Information and add it to the project risk table}
\usage{
create_risk_prog(erisk_orgs, erisk_dist, erisk_msc, risk_item_db)
}
\arguments{
\item{erisk_orgs}{data frame; an organization data frame}

\item{erisk_dist}{data frame; a data frame of districts}

\item{erisk_msc}{data frame; a data frame of mscs}

\item{risk_item_db}{data frame; a data frame of combined projects and risks}
}
\value{
a dataframe of active risks, program, projects ,cost and schedule means
}
\description{
Combine Organization and Division Information and add it to the project risk table
}
\examples{
#Get test data
#erisk_orgs <- erarr::erisk_orgs
#erisk_dist <- erarr::erisk_dist
#erisk_msc <- erarr::erisk_msc
#risk_item_db <- erarr::risk_item_db

#Create items
#risk_project_orgs<-create_risk_prog(erisk_orgs, erisk_dist, erisk_msc, risk_item_db)

}
