% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_riskitem_table.R
\name{create_risk_proj}
\alias{create_risk_proj}
\title{Combine Projects, Means, and Risk Items and only select active risks}
\usage{
create_risk_proj(risk, project, archive)
}
\arguments{
\item{risk}{data frame; a risk item data frame}

\item{project}{data frame; a project data frame}

\item{archive}{data frame; a data frame of modeled and archived risks}
}
\value{
a dataframe of active risks, projects,cost and schedule means
}
\description{
Combine Projects, Means, and Risk Items and only select active risks
}
\examples{
#Get test data
risk <- erarr::erisk_item
project <- erarr::erisk_project
archive <- erarr::erisk_modeled

#Create items
risk_item_db<-create_risk_proj(risk, project, archive)

}
