% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrangleRiskImpact.R
\name{impactcolumn}
\alias{impactcolumn}
\title{Wrangle Risk Impact}
\usage{
impactcolumn(risk_item_table, impacttype)
}
\arguments{
\item{risk_item_table}{data frame; A data frame of a filtered risk item.}
}
\value{
A formatted data table of risk impacts.

#example
milestone_plot<-pieplot_prep(riskdf=risk_item, rankcol)
}
\description{
Conditionally filters Cost and Schedule risk impacts for 
reporting
}
