% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestoneplot.R
\name{milestoneplot}
\alias{milestoneplot}
\title{Milestone Plot}
\usage{
milestoneplot(riskitem, milestonedf)
}
\arguments{
\item{riskitem}{data frame; A data frame of a filtered risk item.}

\item{milestonedf}{data frame; A data frame of phase-milestone pairings}
}
\value{
A ggplot of current milestone on timeline of all project milestones.
}
\description{
Creates a milestone chart
}
\examples{
# Get milestone data and risk item
# milestonedf <- read_csv("inst/app/data/erisk_phase_milestone.csv")
# milestone_df <- data.frame(milestonedf)

# riskitem <- erarr::erisk_item

# milestone_plot <- milestoneplot(riskitem = riskitem, 
#                                 milestonedf = milestone_df)

}
