% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_plots.R
\name{pie_plots}
\alias{pie_plots}
\title{Pie Plots}
\usage{
pie_plots(cost_pie, schedule_pie, perform_pie)
}
\arguments{
\item{cost_pie}{data frame; A data frame of filtered cost risks}

\item{schedule_pie}{data frame;}

\item{perform_pie}{data frame;}
}
\value{
A plotly plot containing 3 pie charts of project risk.
}
\description{
Creates a figure of three pie charts for the specified 
data frames of Cost, Schedule, and Performance risks according to risk 
category ("High", "Medium", "Low", "Opportunity").
}
