% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pieprep.R
\name{pieprep}
\alias{pieprep}
\title{Pie Prep}
\usage{
pieprep(riskdf, rankcol)
}
\arguments{
\item{riskdf}{data frame; A data frame of a filtered risk item.}

\item{rankcol}{character; The column name containing the risk ranks. Ranks
values must be one of: "High", "Medium", "Low", or 
"Opportunity".}
}
\value{
A data.frame of colors for each risk rank.
}
\description{
Preps data for pie charts. Sets the color for each Cost or 
Schedule risk rank.
}
