% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_program_top4s.R
\name{ProgramTop4Report}
\alias{ProgramTop4Report}
\title{Program Top 4 Risks Report}
\arguments{
\item{progID}{numeric; program number}

\item{missionID}{character; mission ID}

\item{MSCID}{character; division ID}

\item{districtID}{character; district ID}

\item{phaseID}{character; phase ID}

\item{disciplineID}{character; discipline id}
}
\description{
Creates a report of all project risks.
}
\details{
All package reports are stored in the folder: `./inst/app/rmd/`
}
