% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_tableprep.R
\name{proj_tableprep}
\alias{proj_tableprep}
\title{Wrangle Project Tables}
\usage{
proj_tableprep(riskdf, colname)
}
\arguments{
\item{projdf}{data frame; A data frame of cost, schedule, and performance risks for projects}

\item{rankcol}{character; The column name containing type of risk, 'Cost", "Schedule', or 'Performance'}
}
\value{
A data.frame of risks summarized for each project
}
\description{
Preps data for table on Project level reports, aggregates cost, schedule, and performance
risks per each project
}
