% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_project_all_risks.R
\name{ProjectAllRiskReport}
\alias{ProjectAllRiskReport}
\title{Project All Risks Report}
\arguments{
\item{projID}{character; erisk Project identifier}

\item{p2ID}{numeric; USACE P2 number}

\item{p2sub}{numeric; P2 sub identifier}
}
\description{
Creates a report of all project risks.
}
\details{
All package reports are stored in the folder: `./inst/app/rmd/`
}
