% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_project_Top4s.R
\name{ProjectTop4s}
\alias{ProjectTop4s}
\title{Project Top 4 Risks Report}
\arguments{
\item{projID}{character; erisk Project identifier}

\item{p2ID}{numeric; USACE P2 number}

\item{p2sub}{numeric; P2 sub identifier}
}
\description{
Creates a report of top 4 risks.
}
\details{
All package reports are stored in the folder: `./inst/app/rmd/`
}
