% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskitemproj.R
\name{riskitemproj}
\alias{riskitemproj}
\title{Risk Item Proj Table}
\usage{
riskitemproj(riskdf, projdf)
}
\arguments{
\item{riskdf}{data frame; A data frame of all risk items.}

\item{proj}{data frame; A data frame of all projects.}
}
\value{
A ggplot of current milestone on timeline of all project milestones.
}
\description{
Creates the Risk Item Project table used to develop project level reports
}
\examples{
riskdf<- read_csv("inst/app/data/RISKLIST_FULL.csv",show_col_types = FALSE)
projdf <- read_csv("inst/app/data/PROJECTLIST_FULL.csv",show_col_types = FALSE)


#example
riskitemproj<-riskitemproj(riskdf, projdf)

}
