% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_risk_item.R
\name{RiskItemReport}
\alias{RiskItemReport}
\title{Risk Item Report}
\arguments{
\item{riskID}{character; erisk Risk identifier}

\item{projID}{character; erisk Project identifier}

\item{p2ID}{numeric; USACE P2 number}
}
\description{
Creates a report of for a specific risk item.
}
\details{
All package reports are stored in the folder: `./inst/app/rmd/`
}
