% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_plots.R
\name{pie_plots}
\alias{pie_plots}
\title{Pie Plots}
\usage{
pie_plots(cost_pie, schedule_pie, perform_pie)
}
\arguments{
\item{cost_pie}{data frame; A data frame of wrangled cost pie data.}

\item{schedule_pie}{data frame; A data frame of wrangled schedule pie data.}

\item{perform_pie}{data frame; A data frame of wrangled performance pie data.}
}
\value{
A plotly plot of pie charts for cost, schedule, and performance risk.
}
\description{
Creates a pie plots from wrangled risk table data
}
\examples{
erisk_item <- read_csv("C:/workspace/e-rarr/RiskItemReportShinyApp/data/RISKLIST_FULL.csv",show_col_types = FALSE)
risk_item_db<-data.frame(erisk_item)
erisk_project <- read_csv("C:/workspace/e-rarr/RiskItemReportShinyApp/data/PROJECTLIST_FULL.csv",show_col_types = FALSE)
risk_project_db<-data.frame(erisk_project)



erisk_ItemProj<-left_join(risk_item_db, risk_project_db, by = "PROJECT_ID")


RiskImpactTable<-risk_item_db[,c("PROJECT_NAME", "RISK_NAME", "USACE_ORGANIZATION","P2_NUMBER")]
riskpies <- erisk_ItemProj |>
  select("P2_NUMBER.x", "RISK_IDENTIFIER","PROJECT_NAME.x","RISK_NAME","RISKCATEGORY",
         "DISCIPLINE", "USACE_ORGANIZATION.x", "COST_RANK_DESC", "SCHEDULE_RANK_DESC", "PERFORMANCE_RANK_DESC")



#example
milestone_plot<-pieplot_prep(riskdf=risk_item, rankcol)

}
